+++ src/OpenTelemetry/BatchExportProcessor.cs
@@ -14,6 +14,8 @@
 // limitations under the License.
 // </copyright>
 
+#nullable enable
+
 using System;
 using System.Diagnostics;
 using System.Threading;
@@ -100,7 +102,13 @@ namespace OpenTelemetry
             {
                 if (this.circularBuffer.Count >= this.maxExportBatchSize)
                 {
-                    this.exportTrigger.Set();
+                    try
+                    {
+                        this.exportTrigger.Set();
+                    }
+                    catch (ObjectDisposedException)
+                    {
+                    }
                 }
 
                 return; // enqueue succeeded
@@ -121,7 +129,14 @@ namespace OpenTelemetry
                 return true; // nothing to flush
             }
 
-            this.exportTrigger.Set();
+            try
+            {
+                this.exportTrigger.Set();
+            }
+            catch (ObjectDisposedException)
+            {
+                return false;
+            }
 
             if (timeoutMilliseconds == 0)
             {
@@ -186,7 +201,15 @@ namespace OpenTelemetry
         protected override bool OnShutdown(int timeoutMilliseconds)
         {
             this.shutdownDrainTarget = this.circularBuffer.AddedCount;
-            this.shutdownTrigger.Set();
+
+            try
+            {
+                this.shutdownTrigger.Set();
+            }
+            catch (ObjectDisposedException)
+            {
+                return false;
+            }
 
             OpenTelemetrySdkEventSource.Log.DroppedExportProcessorItems(this.GetType().Name, this.exporter.GetType().Name, this.droppedCount);